+++ src/network/model/packet-tag-list.cc
@@ -377,7 +377,7 @@ PacketTagList::Serialize(uint32_t* buffer, uint32_t maxSize) const
     }
 
     // Serialized successfully
-    return 1;
+    return size;
 }
 
 uint32_t